requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cart==1.2.1
Django==4.0.5
django-filter==22.1
django-mathfilters==1.0.0
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2-binary==2.9.3
pycryptodome==3.15.0
pytz==2022.1
sqlparse==0.4.2
